Restructure and create first pydantic model
