+++ app/db/create_tables.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 
 from app.db.config import DATABASE_URL
-from app.models.product_listing import create_product_listing
+from app.models.schema import create_product_listing
 
 engine = create_engine(DATABASE_URL, echo=True)
