+++ app/models/pydantic.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class ProductPayloadSchema(BaseModel):
+    pass
